Refactoring/handler structure

+++ src/currency_exchange/dtos.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
@@ -21,27 +22,27 @@ class RateDto:
     id: int
     base_currency: CurrencyDto
     target_currency: CurrencyDto
-    rate: float
+    rate: Decimal
 
 
 @dataclass
 class RatePostUpdateDto:
     base_currency_code: str
     target_currency_code: str
-    rate: float
+    rate: Decimal
 
 
 @dataclass
 class ExchangeDto:
     base_currency: CurrencyDto
     target_currency: CurrencyDto
-    rate: float
-    amount: float
-    converted_amount: float
+    rate: Decimal
+    amount: Decimal
+    converted_amount: Decimal
 
 
 @dataclass
 class ExchangePostDto:
     from_currency_code: str
     to_currency_code: str
-    amount: float
+    amount: Decimal
